refactor function in files
